notes and toast removed.
my notes:
AddNewBook: -inside "onNewIntent ", toast with odd massage
                       -deleted toasts inside catch phrase

AssignBookToShelf: -there is a group of variables with no use, I signed it with                                       "??????", delete it if needed.
                                 -deleted 2 toasts inside async task of update

BookCartActivity: -I didn't touched the toast in GetBookBarcode because I didn't                               know if you finished tests

OnItemClickListenerListView: -I didn't touched it because it is not used. delete
                                                it if you want.

ReturnRouteActivity: - I didn't touched because you may change it.

SearchBook: -I didn't deleted the old url for search, maybe we will use it.

WebActivity: -I didn't touched, delete it if needed.